move raster logic into shared library
