+++ gateware/src/tiliqua/video.py
@@ -1,3 +1,5 @@
+# Utilities for synthesizing digital video timings and presenting a framebuffer.
+#
 # Copyright (c) 2024 Seb Holzapfel, apfelaudio UG <info@apfelaudio.com>
 #
 # SPDX-License-Identifier: CERN-OHL-S-2.0
